chore: extract df-dialog-utils.ts [YTFERONT-3683]
